qwen_graph_rag/qwen_graph_rag/requirements.txt
torch>=2.0.0
torch-geometric
neo4j>=5.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
httpx>=0.24.0
dashscope>=1.14.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.20
chromadb>=0.4.22
huggingface-hub>=0.19.0
sentence-transformers>=2.2.2
